+++ AIStep/src/AIS.Domain/Models/DayStepsInfo.cs
@@ -8,5 +8,5 @@ public class DayStepsInfo
 
     public DateOnly Date { get; set; }
 
-    public long StepsCount { get; set; }
+    public int StepsCount { get; set; }
 }
